Fix insane memory leak and script halt due to Kiota
Trying to write to any of Kiota's stored models (InMemoryStore), including
the additional_data section will make Kiota's memory checking insanely slow to
the point that it effectively halts.

Specifically, the "_ensure_collection_size_is_consistent" in kiota-abstractions
is always running when the script is killed by Ctrl-C.

A Python script taking an upward of 4GB of memory and being killed by the OS is
unacceptable, so this commit fixes it by writing the metadata that we initially
wrote to the additional_data section to a separate JSON file.